Add Playwright UI tests to CI pipeline
